Make input file parsing more robust.
The layout of Rust structures is not guaranteed so using
`mem::size_of()` to detect the type of file isn't reliable. Further,
in each case where `read_base64_file` is used, only one type of file
is acceptable. Therefore, we can have `read_base64_file` return the
base64-decoded file contents and then have the caller decode it as
whatever type of data the caller is expecting.